move tests to schema-examples
